WS3 3.4: Move CLI-resident logic to library crates
- rh-foundation::cli: add parse_package_spec(name@version) and
  expand_home_dir(~/...) utilities
- rh-validator::report: new module with all validation report renderers
  (print_single_result_text/json, print_batch_results_*, print_labeled_*,
  print_operation_outcome, print_batch_operation_outcomes)
- apps/rh-cli/src/validator.rs: replace ~290 lines of local print functions
  with rh_validator::report:: calls; import report module
- apps/rh-cli/src/snapshot.rs: replace local parse_package_spec and
  expand_home_dir with rh_foundation::cli::{parse_package_spec, expand_home_dir}

All 68 rh-cli tests + rh-validator tests pass.
